ContextLock bug fix& Compiler automatically generates timer & more functions added in job manager
1. This is really exciting for me. fixed a bug in ContextLock. This lock caused in some conditions the event is never called up when the previous ones finishes. After the bug, I saw no more deadlocks/mutex incorrect waiting after many tests.

2. Added more compiler support for automatically generating timer for packet retransmission.
Added a sender_timer, and the timer event handler.
Tested to be working fine. Packet retransmission is working properly.

3. Job manager now has more controllability.
I can not only create new remote processes, but also terminate them. I can specify whether I want snapshot or not after termination. The reason for keeping snapshot is of course for doing state/context migration.

+++ application/unit_app/unit_app.cc
@@ -62,8 +62,25 @@ void contextUpdateHandler(int signum){
     loadInitContext( params::get<mace::string>("initcontext")  );
 }
 
+#include "Ticket.h"
+#include "ContextLock.h"
 void snapshotHandler(int signum){
     // TODO: serialize the service class, and store it into a file.
+    //mace::AgentLock lock(mace::AgentLock::WRITE_MODE);
+
+    // get new ticket, but don't use it.
+    uint64_t myTicketNum = Ticket::newTicket();
+    
+    std::cout<<"migration/snapshot ticket="<<myTicketNum<<std::endl;
+    while( true ){
+        SysUtil::sleepu(1000);
+        if( mace::ContextLock::lastCommittedTicket == myTicketNum - 1 ){
+            break;
+        }
+    }
+    // if the thread reaches here, all previous events have been committed, and all late events are blocked.
+    // we can safely take snapshot
+
     mace::Serializable* serv = dynamic_cast<mace::Serializable*>(globalMacedon);
     char tempFileName[] = "ssobj_XXXXXX";
     mace::string buf;
@@ -96,12 +113,12 @@ bool resumeServiceFromFile(mace::Serializable* globalMacedon, mace::string seria
 }
 void shutdownHandler(int signum){
     std::cout<<"received SIGTERM or SIGINT! Ready to stop."<<std::endl;
-    mace::AgentLock lock(mace::AgentLock::WRITE_MODE);
     snapshotHandler(signum);
 
-      std::cout<<" Tell heartbeat proc snapshot is finished"<< std::endl;
+    if( params::get<bool>("killparent",false) == true ){
+      std::cout<<" Tell heartbeat process the snapshot is finished"<< std::endl;
       kill( getppid() , SIGUSR1);
-
+    }
 
       std::cout << "Exiting at time " << TimeUtil::timeu() << std::endl;
